Unify filter redstone behavior
