+++ src/main/resources/assets/gregtech/lang/en_US.lang
@@ -309,7 +309,8 @@ GT5U.machines.advdebugstructurewriter.gui.transpose.tooltip=Transpose
 GT5U.machines.nei_transfer.steam.tooltip=%s steam recipes
 GT5U.machines.nei_transfer.voltage.tooltip=Recipes available in %s
 GT5U.machines.emit_energy.tooltip=Emit energy to output side
-GT5U.machines.emit_redstone.tooltip=Emit redstone if no slot is free
+GT5U.machines.emit_redstone_if_full.tooltip=Emit redstone if no slot is free
+GT5U.machines.emit_redstone_gradually.tooltip=Emit redstone for each slot in use
 GT5U.machines.invert_redstone.tooltip=Invert redstone
 GT5U.machines.buffer_stocking_mode.tooltip=Enable stocking mode
 GT5U.machines.buffer_stocking_mode.tooltip.extended=§7Keeps this many items in destination rather than transfer items in batches of this amount.
